implement pin api client and inventory mutations
Add PIN API integration with typed failure handling and introduce mutable ATM inventory primitives so withdrawal processing can safely update note counts.

Made-with: Cursor

+++ packages/shared/src/schemas/atm.ts
@@ -8,7 +8,7 @@ export const pinRequestSchema = z.object({
 });
 
 export const pinApiSuccessResponseSchema = z.object({
-  currentBalance: z.number().int(),
+  currentBalance: z.number(),
 });
 
 export const noteCountsSchema = z.object({
